seperate category component from  ecommerce content
