+++ frontend/warehouse-app/src/app/ecommerce/page.tsx
@@ -1,6 +1,4 @@
 "use client";
-
-import React from "react";
 import EcommerceContent from "@/components/ecommerce/EcommerceContent";
 
 export default function Ecommerce() {
