+++ frontend/warehouse-app/src/components/Category/Category.tsx
@@ -1,24 +1,42 @@
-import { useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import {useMediaQuery, useTheme } from "@mui/material";
-import { useGridSkeletonCount } from "@/hooks/useGridSkeletonCount";
 import { debounce } from "@/utils/debounce";
 import CategorySection from "../ecommerce/category_temp/CategorySection";
+import useCategoryStore from "@/store/categoryStore";
+import { useDetectUserLocation } from "@/hooks/useDetectUserLocation";
 
-const Category = () => {
+//TODO P0: Review this component and recorrect it properly
+const Category = ({ slug }: { slug: string | undefined }) => {
+    const { countryCode } = useDetectUserLocation();
     const theme = useTheme();
     const skeletonRef = useRef<HTMLDivElement | null>(null);
     const [rowHeight, setRowHeight] = useState(300);
     const isSm = useMediaQuery(theme.breakpoints.up('sm'));
 
     const [categoryBottom, setCategoryBottom] = useState(100);
-    const skeletonCount = useGridSkeletonCount({
-        itemHeight: rowHeight,
-        offsetY: categoryBottom + 32,
-        minCount: 4
-    });
+    const { getCategories, selectedCategory, setCategory } = useCategoryStore();
+
+    const syncCategoryWithSlug = useCallback(() => {
+        const nextCategory = slug ?? null;
+        if (nextCategory !== selectedCategory) {
+          setCategory(nextCategory);
+        }
+      }, [slug, selectedCategory, setCategory]);
 
     useEffect(() => {
+    syncCategoryWithSlug();
+    }, []);
 
+    const fetchCategoryList = useCallback(() => {
+    if (!countryCode) return [];
+    getCategories(countryCode);
+    }, [countryCode, getCategories]);
+
+    useEffect(() => {
+        fetchCategoryList();
+    }, [fetchCategoryList]);
+
+    useEffect(() => {
         const measureLayout = () => {
             if (skeletonRef.current) {
                 const cardHeight = skeletonRef.current.getBoundingClientRect().height;
